Course list and create. Done

+++ StudyONU.Core/Entities/CourseEntity.cs
@@ -11,7 +11,7 @@ namespace StudyONU.Core.Entities
 
         public DateTime DateCreated { get; set; }
 
-        public int CourseNumber { get; set; }
+        public byte CourseNumber { get; set; }
 
         public bool IsPublished { get; set; }
