+++ StudyONU.Logic/DTO/Course/CourseCreateDTO.cs
@@ -4,7 +4,7 @@
     {
         public string Name { get; set; }
 
-        public int CourseNumber { get; set; }
+        public byte CourseNumber { get; set; }
 
         public bool IsPublished { get; set; }
